# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sound2Soul API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /creators/{slug}/tracks:
    get:
      operationId: getCreatorTracks
      tags: [creators]
      summary: Get public tracks by creator slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"

  # Tracks
  /tracks:
    get:
      operationId: getMyTracks
      tags: [tracks]
      summary: Get creator's own tracks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
    post:
      operationId: createTrack
      tags: [tracks]
      summary: Upload a new track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tracks/{slug}:
    get:
      operationId: getTrackBySlug
      tags: [tracks]
      summary: Get a public track by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTrack
      tags: [tracks]
      summary: Update own track
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
    delete:
      operationId: deleteTrack
      tags: [tracks]
      summary: Delete own track
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tracks/{slug}/play:
    post:
      operationId: incrementPlayCount
      tags: [tracks]
      summary: Increment play count
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Explore
...
  /upload/audio-url:
    post:
      operationId: getAudioUploadUrl
      tags: [tracks]
      summary: Get a pre-signed URL for audio upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /upload/image-url:
    post:
      operationId: getImageUploadUrl
      tags: [tracks]
      summary: Get a pre-signed URL for image upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

  # Admin
...
  /admin/tracks:
    get:
      operationId: getAdminTracks
      tags: [admin]
      summary: Get all tracks for admin
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTrack"
...
  /admin/tracks/{trackId}:
    delete:
      operationId: adminDeleteTrack
      tags: [admin]
      summary: Admin delete a track
      parameters:
        - name: trackId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    CreatorPublicProfile:
      type: object
      properties:
        id:
          type: string
        artistName:
          type: string
        slug:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        bannerUrl:
          type: string
        genres:
          type: array
          items:
            type: string
        moodIdentityTags:
          type: array
          items:
            type: string
        aiToolsUsed:
          type: array
          items:
            type: string
        socialLinks:
          type: object
          additionalProperties:
            type: string
        creatorStatement:
          type: string
        followerCount:
          type: integer
        trackCount:
          type: integer
        createdAt:
          type: string
      required: [id, artistName, slug]
...
    Track:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
        moderationStatus:
          type: string
          enum: [pending, approved, rejected]
        isFeatured:
          type: boolean
        playCount:
          type: integer
        likeCount:
          type: integer
        saveCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        creator:
          $ref: "#/components/schemas/CreatorPublicProfile"
      required: [id, creatorId, title, slug, visibility, moderationStatus]
    TrackDetail:
      type: object
      properties:
        id:
          type: string
        creatorId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
        moderationStatus:
          type: string
          enum: [pending, approved, rejected]
        isFeatured:
          type: boolean
        playCount:
          type: integer
        likeCount:
          type: integer
        saveCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        creator:
          $ref: "#/components/schemas/CreatorPublicProfile"
        isLiked:
          type: boolean
        isSaved:
          type: boolean
      required: [id, creatorId, title, slug, visibility, moderationStatus]
    CreateTrackBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
      required: [title, audioUrl, visibility]
    UpdateTrackBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        audioUrl:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        moodTags:
          type: array
          items:
            type: string
        soulStory:
          type: string
        aiInvolvementType:
          type: string
        humanContributionChecklist:
          type: object
        rightsConfirmation:
          type: object
        visibility:
          type: string
          enum: [draft, public, unlisted]
...
    UploadUrlBody:
      type: object
      properties:
        filename:
          type: string
        contentType:
          type: string
      required: [filename, contentType]
    UploadUrlResponse:
      type: object
      properties:
        uploadUrl:
          type: string
        publicUrl:
          type: string
      required: [uploadUrl, publicUrl]
...
    AdminTrack:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        coverImageUrl:
          type: string
        genre:
          type: string
        aiInvolvementType:
          type: string
        visibility:
          type: string
        moderationStatus:
          type: string
        isFeatured:
          type: boolean
        playCount:
          type: integer
        likeCount:
          type: integer
        reportCount:
          type: integer
        createdAt:
          type: string
        creator:
          $ref: "#/components/schemas/CreatorPublicProfile"
      required: [id, title, slug, visibility, moderationStatus]
...